Cambios implementados:
1. Visulización de puntaje en la lista de lecciones
2. Notificación de perdida de progreso al salir de la leccion
3. Ajuste en los puntajes al presionar volver a jugar o rehacer una leccion ya completada
4. Persistencia de puntos por usuario

+++ app/src/main/java/com/unam/algeplus/model/Termino.java
@@ -10,12 +10,12 @@ public class Termino {
     public enum TipoTermino { VARIABLE, CONSTANTE }
     public enum LadoEcuacion { IZQUIERDO, DERECHO }
 
-    private double valor;
+    private float valor;
     private TipoTermino tipo;
     private int signo;          // +1 o -1
     private LadoEcuacion lado;
 
-    public Termino(double valor, TipoTermino tipo, int signo, LadoEcuacion lado) {
+    public Termino(float valor, TipoTermino tipo, int signo, LadoEcuacion lado) {
         this.valor = valor;
         this.tipo = tipo;
         this.signo = signo;
@@ -23,7 +23,7 @@ public class Termino {
     }
 
     public double getValor() { return valor; }
-    public void setValor(double valor) { this.valor = valor; }
+    public void setValor(float valor) { this.valor = valor; }
 
     public TipoTermino getTipo() { return tipo; }
     public int getSigno() { return signo; }
